Added search functionality for "by Content". UI changes. Create User link for the Admin User. "Users" screen also gives option to "Grant" or "Revoke" admin permissions. CSS for all the buttons!

+++ app/controllers/posts_controller.rb
@@ -16,17 +16,32 @@ class PostsController < ApplicationController
   # GET /posts/1
   # GET /posts/1.json
   def show
-    if session[:id]
-      @post = Post.find(params[:id])
-      @posts = Post.get_replies(params[:id])
-
-      respond_to do |format|
-        format.html # show.html.erb
-        format.json { render json: @post }
-      end
-    else
+    if params[:from_reply_button] == '1' and !session[:id]
       flash[:error] = "You must Sign in to Post a reply"
       redirect_to :controller => 'system', :action => 'index'
+    else
+      @post = Post.find(params[:id])
+
+      if @post
+        if !@post.parent.nil?
+          @post = Post.find(@post.parent)
+        end
+
+        if @post
+          @posts = Post.get_replies(@post.id)
+
+          respond_to do |format|
+            format.html # show.html.erb
+            format.json { render json: @post }
+          end
+        else
+          flash[:error] = "The post you are trying to view does not exist."
+          redirect_to :controller => 'system', :action => 'index'
+        end
+      else
+        flash[:error] = "The post you are trying to view does not exist."
+        redirect_to :controller => 'system', :action => 'index'
+      end
     end
   end
 
@@ -48,7 +63,7 @@ class PostsController < ApplicationController
 
   # GET /posts/1/edit
   def edit
-    @post = Post.find(params[:id])
+    redirect_to :controller => 'system', :action => 'index'
   end
 
   # POST /posts
@@ -70,6 +85,7 @@ class PostsController < ApplicationController
   # PUT /posts/1
   # PUT /posts/1.json
   def update
+    redirect_to :controller => 'system', :action => 'index'
     @post = Post.find(params[:id])
 
     respond_to do |format|
@@ -87,23 +103,31 @@ class PostsController < ApplicationController
   # DELETE /posts/1.json
   def destroy
 
-    @myreplies = Post.find_all_by_parent(params[:id])
-    if !@myreplies.nil?
-      @myreplies.each do |myreply|
-        delete_my_votes(myreply.id)
-        myreply.destroy
+    if session[:id] and session[:isadmin] == 1
+      @myreplies = Post.find_all_by_parent(params[:id])
+      if !@myreplies.nil?
+        @myreplies.each do |myreply|
+          delete_my_votes(myreply.id)
+          myreply.destroy
+        end
       end
+
+      @post = Post.find(params[:id])
+      if @post
+        delete_my_votes(@post.id)
+        @post.destroy
+
+        respond_to do |format|
+          format.html { redirect_to ('/askuery') }
+          format.json { head :ok }
+        end
+      else
+        flash[:error] = "The post is already deleted"
+        redirect_to :controller => 'system', :action => 'index'
+      end
+    else
+      flash[:error] = "You must be signed in as an Administrator to perform this action."
+      redirect_to :controller => 'system', :action => 'index'
     end
-
-    @post = Post.find(params[:id])
-    delete_my_votes(@post.id)
-    @post.destroy
-
-    respond_to do |format|
-      format.html { redirect_to ('/askuery') }
-      format.json { head :ok }
-    end
-
-    #redirect_to :controller => 'system', :action => 'index'
   end
 end
